Added legend
